pyproject.toml
[project]
name = "kitchen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hoanghapham", email = "hoangha.20021992@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "gradio>=5.35.0",
    "gradio-image-annotation>=0.4.0",
    "ipykernel>=6.29.5",
    "lap>=0.5.12",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "shapely>=2.1.1",
    "ultralytics>=8.3.158",
    "watchdog>=6.0.0",
]

[project.scripts]
kitchen = "kitchen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
